Refactoring) create read.h and renamed filename like write
