+++ SSD/SSD.vcxproj.filters
@@ -27,7 +27,7 @@
     <ClCompile Include="main.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="SSD_Read.cpp">
+    <ClCompile Include="read.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
     <ClCompile Include="write.cpp">
@@ -39,6 +39,9 @@
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="write.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="read.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
